chore(wmg): style cleanup
