+++ frontend/src/views/WheresMyGeneV2/components/Main/style.ts
@@ -1,5 +1,6 @@
 import RightSideBar from "src/components/common/RightSideBar";
 import styled from "@emotion/styled";
+import { Drawer } from "@blueprintjs/core";
 
 export const StyledRightSideBar = styled(RightSideBar)`
   /**
@@ -7,4 +8,16 @@ export const StyledRightSideBar = styled(RightSideBar)`
    * Gene Search Bar when the viewport width is narrow enough that they overlap
    */
   z-index: 3;
+`;
+
+export const StyledSidebarDrawer = styled(Drawer)`
+  .bp4-drawer-header {
+    box-shadow: none;
+  }
+`;
+
+export const StyledBannerContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  padding-bottom: 20px;
 `;
