--- frontend/src/views/WheresMyGeneV2/components/Main/style_old.ts
@@ -1,14 +1,0 @@
-import { Drawer } from "@blueprintjs/core";
-import styled from "@emotion/styled";
-
-export const StyledSidebarDrawer = styled(Drawer)`
-  .bp4-drawer-header {
-    box-shadow: none;
-  }
-`;
-
-export const StyledBannerContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  padding-bottom: 20px;
-`;
