+++ frontend/src/views/WheresMyGeneV2/style.ts
@@ -7,7 +7,7 @@ import { FOOTER_HEIGHT_PX } from "src/components/Footer/style";
 import { HEADER_HEIGHT_PX } from "src/components/Header/style";
 import { SidebarMainWrapper } from "src/components/Layout/style";
 import { Container } from "./components/GeneSearchBar/style";
-import { LegendWrapper } from "./components/InfoPanel/components/Legend/style";
+import { LegendWrapper } from "../WheresMyGeneV2/components/InfoPanel/components/Legend/style";
 
 export const LEGEND_MARGIN_BOTTOM_PX = 20;
 
@@ -24,10 +24,10 @@ export const Gap = styled.div`
 
 export const Top = styled.div`
   display: flex;
-  justify-content: space-between;
-  margin-bottom: ${LEGEND_MARGIN_BOTTOM_PX}px;
+  justify-content: end;
 
   /* The CSS below is to center the legend in the PNG output */
+
   &.CLONED {
     justify-content: center;
     margin: 0;
